requirements.txt
flask
huggingface_hub
mysql-connector-python
boto3
tabulate
flask_cors
